refactoring game scene
